+++ src/main/java/com/bookripple/api/domain/order/code/TradeErrorCode.java
@@ -1,7 +1,7 @@
 package com.bookripple.api.domain.order.code;
 
 
-import com.bookripple.api.common.code.BaseErrorCode;
+import com.bookripple.api.global.code.BaseErrorCode;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
